+++ src/components/App.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import Game from './Game';
+
+class App extends React.Component {
+  render() {
+    return(
+      <div>
+        <Game />
+      </div>
+    );
+  }
+}
+
+export default App;
